Fixed configuration test errors on Visual Studio.
Fixed inappropriate use of CHECK_FUNCTION_EXISTS() for detecting 'memcmp',
'memcpy' and 'memset', which became apparent by the recent modification that
changed the configuration tests to be run in 'Release' mode.
As described by the CMake documentation, CHECK_FUNCTION_EXISTS() cannot be
used to detect functions that are inlined in headers or specified as a macro
and fails to detect anything in the 32-bit versions of the Win32 API, because
of a mismatch in calling conventions.