Fixed compiling with Oracle Developer Studio (SunPro).
Fixed linker issues caused by missing 'const' modifiers on some by-value
arguments. The compiler seems to understand that these functions are
compatible but then creates symbol references to the spelling that doesn't
exist, leading to linker errors later in the build process.
Removed some of the previously introduced workarounds that are not necessary
on newer versions of SunPro, the old versions that required them lack some
fundamental C++ features that are necessary for building DCMTK anyway (also
removing the comment that SunPro was the retarded child of a Java compiler and
an actual C++ compiler that is no longer true - SunPro can no longer be
considered a child). Use Oracle Developer Studio 12.5 or later!