Small fixes for a previous commit.
Fixed a typo and removed an inappropriate restriction (only affecting the
CMake GUI):
DCMTK_ENABLE_LFS was falsely restricted to the modes that were detected as
available, making it impossible to choose 'lfs' in case 'lfs64' was available
(since in that case the test looking for 'lfs' was not run and it was
therefore unknown whether it is available or not). Now it is also possible to
choose modes where the configuration tests have not been executed yet to
trigger launching the tests in the first place.